pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atlas-vla"
version = "0.1.0"
description = "Vision-Language-Action model based on VGGT for robot manipulation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Tingting Du", email = "tdu35@wisc.edu"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "numpy>=1.24.0",
    "einops>=0.6.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
wandb = [
    "wandb>=0.15.0",
]

[tool.setuptools.packages.find]
exclude = ["tests*", "*.tests*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']
